smart_warehouse_ai/requirements.txt
fastapi
uvicorn[standard]
pandas
sqlalchemy
psycopg2-binary
scikit-learn
joblib
numpy
python-dotenv
